backend/requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
pandas
PyMuPDF
python-docx
python-multipart
requests
boto3